apps/ai-engine/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "supplygraph-ai-engine"
version = "0.1.0"
description = "AI engine for SupplyGraph procurement automation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=1.0.0",
    "psycopg[binary]>=3.1.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "pydantic>=2.5.0",
    "prisma>=0.15.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0", # Keep for fallback compatibility
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",
    "httpx>=0.25.0",
    "websockets>=12.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.31.0",
    "requests>=2.32.5",
    "pyjwt>=2.8.0",
    "stripe>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "I", "N", "B", "C90", "UP"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
